Refactor User entity and add batch user upload feature
The User entity has been converted from a record to a class and various Lombok annotations have been added to simplify the code. A feature has also been added to the UserController to support batch upload of users via a JSON file. This involves parsing the JSON into User objects, saving them to the database, and providing a response with the number of users that were successfully saved and those that failed.